requirements.txt
Django==5.1.2
pycryptodome==3.21.0
